Enable NSGs shared between multiple machines
This commit allows users of the Azure driver to submit an NSG name or ARM resource identifier to be used by multiple machines that are provisioned via Rancher Machine. If a user does not submit an NSG name, machine will default to the legacy option of creating an NSG per machine.

Major updates:
- Azure driver now takes in d.NSG, which is passed into resolveNSGReference to output d.nsgResourceIdentifier (used to find out what resource needs to be created / updated) and d.nsgUsedInPool (used to handle tagging + cleanup logic).
- Introduced securityrules.go to handle the bulk of the security rule logic, including syncing open ports, adding required allow inbound azure security rules, and erroring out / logging any concerning custom ports set by the user in the Azure UI (outside of those set in d.openPorts)
- Introduced tag.go to handle tagging logic and updated NSG cleanup logic to account for tagging

Minor updates:
- Renamed CreateNetworkSecurityGroup to ConfigureNetworkSecurityGroup